move around dependencies, refactor repl
