--- misc/ssh-programmer/requirements.txt
@@ -1,2 +1,0 @@
-paramiko==2.9.2
-keystone-engine==0.9.2
