+++ patches/277-tools/config
@@ -2,4 +2,4 @@ NAME="277-tools"
 BINARY_NAME="ilo4_277.bin"
 BINARY_URL="https://downloads.hpe.com/pub/softlib2/software1/sc-linux-fw-ilo/p192122427/v188589/CP046020.scexe"
 BINARY_SHA1="128f2ce1834dd68e0b2d806243e14fcc27b51364"
-RESULT_SHA1="952dd04199698f861738759b2986323472ce3685"
+RESULT_SHA1="b274fedc2a9cd6e45fe9c238f0f6b9a65144ab12"
