+++ requirements.txt
@@ -1,1 +1,2 @@
 keystone-engine==0.9.2
+paramiko==2.9.2
